+++ source/renderer/app/containers/staking/Staking.js
@@ -76,8 +76,14 @@ export default class Staking extends Component<Props> {
       stores: { app, staking, networkStatus, uiDialogs },
       children,
     } = this.props;
-    const { isSynced, syncPercentage } = networkStatus;
+    const {
+      isSynced,
+      syncPercentage,
+      isAlonzoPending,
+      isAlonzoActivated,
+    } = networkStatus;
     const { isStakingDelegationCountdown } = staking;
+    const shouldShowInfoTab = isAlonzoPending || isAlonzoActivated;
 
     const isDelegationWizardOpen = uiDialogs.isOpen(
       DelegationSetupWizardDialog
@@ -106,7 +112,7 @@ export default class Staking extends Component<Props> {
             isActiveNavItem={this.isActiveNavItem}
             onNavItemClick={this.handleNavItemClick}
             activeItem={app.currentPage}
-            showInfoTab={IS_STAKING_INFO_PAGE_AVAILABLE}
+            showInfoTab={IS_STAKING_INFO_PAGE_AVAILABLE && shouldShowInfoTab}
           >
             {children}
           </StakingWithNavigation>