+++ source/renderer/app/Routes.js
@@ -30,6 +30,7 @@ import WalletSummaryPage from './containers/wallet/WalletSummaryPage';
 import WalletSendPage from './containers/wallet/WalletSendPage';
 import WalletReceivePage from './containers/wallet/WalletReceivePage';
 import WalletTransactionsPage from './containers/wallet/WalletTransactionsPage';
+import WalletTokensPage from './containers/wallet/WalletTokensPage';
 import WalletSettingsPage from './containers/wallet/WalletSettingsPage';
 import WalletUtxoPage from './containers/wallet/WalletUtxoPage';
 import VotingRegistrationPage from './containers/voting/VotingRegistrationPage';
@@ -64,14 +65,15 @@ export const Routes = withRouter(() => (
               path={ROUTES.WALLETS.SUMMARY}
               component={WalletSummaryPage}
             />
-            <Route
-              path={ROUTES.WALLETS.TRANSACTIONS}
-              component={WalletTransactionsPage}
-            />
             <Route path={ROUTES.WALLETS.SEND} component={WalletSendPage} />
             <Route
               path={ROUTES.WALLETS.RECEIVE}
               component={WalletReceivePage}
+            />
+            <Route path={ROUTES.WALLETS.TOKENS} component={WalletTokensPage} />
+            <Route
+              path={ROUTES.WALLETS.TRANSACTIONS}
+              component={WalletTransactionsPage}
             />
             <Route
               path={ROUTES.WALLETS.SETTINGS}