+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Remove Alonzo tada icon and countdown screen ([PR 2708](https://github.com/input-output-hk/daedalus/pull/2708))
 
+### Fixes
+
+- Fixed transaction timestamps localization ([PR 2702](https://github.com/input-output-hk/daedalus/pull/2702))
+
 ## 4.4.0
 
 ### Features